+++ bundler/spec/dependabot/bundler/file_updater_spec.rb
@@ -614,12 +614,12 @@ RSpec.describe Dependabot::Bundler::FileUpdater do
 
           %w(business prius uk_phone_numbers).each do |dep|
             original_remote_line =
-              old_lock.find { |l| l.include?("gocardless/#{dep}") }
+              old_lock.find { |l| l.include?("dependabot-fixtures/#{dep}") }
             original_revision_line =
               old_lock[old_lock.find_index(original_remote_line) + 1]
 
             new_remote_line =
-              new_lock.find { |l| l.include?("gocardless/#{dep}") }
+              new_lock.find { |l| l.include?("dependabot-fixtures/#{dep}") }
             new_revision_line =
               new_lock[new_lock.find_index(original_remote_line) + 1]
 
@@ -638,12 +638,12 @@ RSpec.describe Dependabot::Bundler::FileUpdater do
             new_lock = file.content.split(/^/)
 
             original_remote_line =
-              old_lock.find { |l| l.include?("gocardless/business") }
+              old_lock.find { |l| l.include?("dependabot-fixtures/business") }
             original_revision_line =
               old_lock[old_lock.find_index(original_remote_line) + 1]
 
             new_remote_line =
-              new_lock.find { |l| l.include?("gocardless/business") }
+              new_lock.find { |l| l.include?("dependabot-fixtures/business") }
 
             new_revision_line =
               new_lock[new_lock.find_index(original_remote_line) + 1]
@@ -673,12 +673,12 @@ RSpec.describe Dependabot::Bundler::FileUpdater do
 
             %w(business prius uk_phone_numbers).each do |dep|
               original_remote_line =
-                old_lock.find { |l| l.include?("gocardless/#{dep}") }
+                old_lock.find { |l| l.include?("dependabot-fixtures/#{dep}") }
               original_revision_line =
                 old_lock[old_lock.find_index(original_remote_line) + 1]
 
               new_remote_line =
-                new_lock.find { |l| l.include?("gocardless/#{dep}") }
+                new_lock.find { |l| l.include?("dependabot-fixtures/#{dep}") }
               new_revision_line =
                 new_lock[new_lock.find_index(original_remote_line) + 1]
 
@@ -702,7 +702,7 @@ RSpec.describe Dependabot::Bundler::FileUpdater do
           it "generates the correct lockfile" do
             expect(file.content).to include("statesman (2.0.1)")
             expect(file.content).
-              to include "remote: http://github.com/gocardless/uk_phone_numbers"
+              to include "remote: http://github.com/dependabot-fixtures/uk_phone_numbers"
           end
         end
       end
@@ -726,7 +726,7 @@ RSpec.describe Dependabot::Bundler::FileUpdater do
             groups: [],
             source: {
               type: "git",
-              url: "https://github.com/gocardless/prius",
+              url: "https://github.com/dependabot-fixtures/prius",
               branch: "master",
               ref: "master"
             }
@@ -739,7 +739,7 @@ RSpec.describe Dependabot::Bundler::FileUpdater do
             groups: [],
             source: {
               type: "git",
-              url: "https://github.com/gocardless/prius",
+              url: "https://github.com/dependabot-fixtures/prius",
               branch: "master",
               ref: "master"
             }
@@ -751,12 +751,12 @@ RSpec.describe Dependabot::Bundler::FileUpdater do
           new_lock = file.content.split(/^/)
 
           original_remote_line =
-            old_lock.find { |l| l.include?("gocardless/prius") }
+            old_lock.find { |l| l.include?("dependabot-fixtures/prius") }
           original_revision_line =
             old_lock[old_lock.find_index(original_remote_line) + 1]
 
           new_remote_line =
-            new_lock.find { |l| l.include?("gocardless/prius") }
+            new_lock.find { |l| l.include?("dependabot-fixtures/prius") }
           new_revision_line =
             new_lock[new_lock.find_index(original_remote_line) + 1]
 
@@ -786,7 +786,7 @@ RSpec.describe Dependabot::Bundler::FileUpdater do
               groups: [],
               source: {
                 type: "git",
-                url: "https://github.com/elastic/elasticsearch-ruby.git",
+                url: "https://github.com/dependabot-fixtures/elasticsearch-ruby.git",
                 branch: "5.x",
                 ref: "5.x"
               }
@@ -834,7 +834,7 @@ RSpec.describe Dependabot::Bundler::FileUpdater do
               groups: [],
               source: {
                 type: "git",
-                url: "https://github.com/elastic/elasticsearch-ruby.git",
+                url: "https://github.com/dependabot-fixtures/elasticsearch-ruby.git",
                 branch: "5.x",
                 ref: "5.x"
               }